mlpce: Add global about() function with information about the build
For this to be useful, revision.h needs to be generated with repository
project and revision information. MLPCE_REVISION_HEADER_ENABLED cmake
variable is used to activate the include of revision.h. It is turned
off by default.

Also add luaopen_tosbindl to loadedlibs when MLPCE_TOSBINDL_ENABLED
is defined.